+++ elements/html5.spec.ts
@@ -95,6 +95,7 @@ const tagNames = [
 	"output",
 	"p",
 	"param",
+	"picture",
 	"plaintext",
 	"pre",
 	"progress",