+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Upcoming release
 
+- allow `crossorigin` attribute to be boolean or `""` (maps to `"anonymous"`).
+- add `<picture>` element.
+- mark `<style>` as foreign as to not trigger errors inside css content.
+- allow whitespace around attribute equals sign, e.g `class = "foo"`.
+
 ## 0.25.1 (2019-05-10)
 
 - allow raw ampersands (`&`) in quoted attributes.